[ADDED]    the ghost step functionality was added    Unit tests were added

+++ webapp/users_views.py
@@ -45,7 +45,7 @@ def AddUser():
 @role_required('editor', 'admin')
 def AddUserFromManager():
     if request.method == 'GET':
-        return render_template('add_user_from_manager.jinja2')
+        return render_template('add_user_from_manager.jinja2', user=current_user)
 
     # if it is a POST
     username = request.form['username']
@@ -58,13 +58,12 @@ def AddUserFromManager():
         email
     )
     if not add_user_command.executed:
+
         return render_template(
-            'add_user.jinja2',
-            error_message=add_user_command.error,
-            user=current_user,
-            projects=projects.return_all_projects().data,
-            current_project_id=int(session.get('current_project_id'))
-        )
+            'add_user_from_manager.jinja2',
+            error_message=f"could not create an acount for {username}. error: {add_user_command.error}",
+            user=current_user
+        ), 400
 
     return redirect(url_for('users_views.UsersManagement'))
 
@@ -94,6 +93,9 @@ def UsersManagement():
 
     # if it is a POST
     user_id = request.form['id']
+    if current_user.user_level != 'admin':
+        abort(403)
+
     if request.form['action'] == "deletion":
         command = current_app.Users_manipulation.delete_user(user_id)
         if not command.executed:
@@ -105,39 +107,41 @@ def UsersManagement():
                 user=current_user,
                 projects=projects.return_all_projects().data,
                 current_project_id=int(session.get('current_project_id'))
-            )
+            ), 500
         return redirect(url_for('users_views.UsersManagement'))
 
     if request.form['action'] == "reseting":
-        return redirect(url_for('users_views.ResetUserPassword'))
+        # return redirect(url_for('users_views.ResetUserPassword'))
+        # The redirect is handled by the html. future fix
+        pass
 
     if request.form['action'] == "change_level":
         user_id = request.form['id']
         user_level = request.form['new_level']
-        if not current_app.Users_manipulation.change_user_level(user_id, user_level):
+
+        command = current_app.Users_manipulation.change_user_level(
+            user_id,
+            user_level
+        )
+        if not command.executed:
             return render_template(
                 "error_message.jinja2",
-                error_message="There must be at least one admin user!",
+                error_message=command.error,
                 page_url="users_views.UsersManagement",
                 page_name="Users management",
                 user=current_user,
                 projects=projects.return_all_projects().data,
                 current_project_id=int(session.get('current_project_id'))
             )
-            # return render_template("users_management.jinja2", users=current_app.Users_manipulation.return_users().data, error_message="There must be at least 1 Admin user.")
 
         return redirect(url_for('users_views.UsersManagement'))
 
-    # currently, there is not post allowed for non admin users.
-    if current_user.user_level != 'admin':
-        abort(403)
-
     # In case a non mapped post is sent
-    return render_template('404.jinja2')
+    return render_template('404.jinja2'), 404
 
 
 @users_views.route('/ResetUserPassword/<string:username>', methods=['GET', 'POST'])
-@role_required('admin')
+@role_required('admin', 'editor', 'viewer')
 @login_required
 def ResetUserPassword(username):
     if request.method == 'GET':
@@ -148,7 +152,7 @@ def ResetUserPassword(username):
             current_project_id=int(session.get('current_project_id'))
         )
 
-    if current_user.user_level != 'admin':
+    if current_user.user_level != 'admin' and username != current_user.username:
         abort(403)
     if request.form['action'] == "reseter":
         password = request.form['newPassword']
